#8 refactoring
