--- src/components/Parameters.js
@@ -1,12 +1,0 @@
-import React, { Component } from 'react';
-import '../style/parameters.css';
-
-class Parameters extends Component {
-  render () {
-    return (
-      <div className='parameters' />
-    );
-  }
-}
-
-export default Parameters;
