--- src/pages/Favorites.js
@@ -1,12 +1,0 @@
-import React, { Component } from 'react';
-import '../style/favorites.css';
-
-class Favorites extends Component {
-  render () {
-    return (
-      <div className='favorites' />
-    );
-  }
-}
-
-export default Favorites;
