--- src/style/cardsWeatherWears.css
@@ -1,58 +1,0 @@
-.weatherCard {
-    box-shadow: 0px 70px 40px -50px rgba(0, 0, 0, 0.2), 0px 3px 20px rgba(0, 0, 0, 0.8);
-    margin: auto;
-    width: 70%;
-    background-image: linear-gradient(135deg, #8bc6ec63 0%, #95d0e2b0 100%);
-    text-align: center;
-    margin-top: 60px;
-}
-
-.dontForget {
-    background-color: #ffffff47;
-    width: 100%;
-    margin-bottom: 0;
-    padding: 10px;
-    color: white;
-    border-top: solid #8080807d;
-}
-
-.humanoid {
-    width: 60%;
-    border-radius: 10px;
-}
-
-.clothes {
-    padding: 15px;
-    display: flex;
-    background-color: #fff;
-    justify-content: space-around;
-}
-
-.clothes img {
-    width: 14%;
-}
-
-.ourSuggest, .textClothes {
-    padding: 20px;
-}
-
-.ourSuggest {
-    color: #073a54;
-    font-weight: bold;
-}
-
-.textClothes {
-    color: #005C89;
-}
-
-@media only screen and (max-width: 800px) {
-
-    .humanoid {
-        width: 40%;
-    }
-
-    .clothes img {
-        width: 20%;
-    }
-
-}
